renamed src directory to package name python project structure normally does not have a src folder
